# webconsole-back/swagger/SGDSI_Website4sgCore_API_ADMIN_PROD.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Website4sg Core API ADMIN PROD documentation",
    "version" : "1.3",
    "title" : "SGDSI Website4sg Core API ADMIN PROD"
  },
  "host" : "core.website4sg.saint-gobain.com",
  "basePath" : "/",
...
  "schemes" : [ "https" ],
  "paths" : {
    "/tools/incapsula/sites" : {
      "post" : {
        "tags" : [ "toolsincapsulasites" ],
        "summary" : "Add a site in Incapsula",
        "description" : "Add a site in Incapsula",
        "operationId" : "addWebsite",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Site",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/WAF Site"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Operation successfully done.",
            "schema" : {
              "$ref" : "#/definitions/IncapsulaResponse"
            }
          },
          "400" : {
            "description" : "Invalid data supplied"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
...
    "/tools/incapsula/sites/{domainCode}/certificate" : {
      "delete" : {
        "tags" : [ "toolsincapsulasites" ],
        "summary" : "Remove the certificate of the site from Incapsula",
        "description" : "Remove the certificate of the site from Incapsula",
        "operationId" : "removeCertificate",
        "parameters" : [ {
          "name" : "domainCode",
          "in" : "path",
          "description" : "Domain Id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Operation successfully done.",
            "schema" : {
              "$ref" : "#/definitions/IncapsulaResponse"
            }
          },
          "400" : {
            "description" : "Invalid data supplied"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
    "/tools/incapsula/sites/{domainCode}/certificate/{certificateCode}" : {
      "put" : {
        "tags" : [ "toolsincapsulasites" ],
        "summary" : "Upload the certificate to the site in Incapsula",
        "description" : "Upload the certificate to the site in Incapsula",
        "operationId" : "uploadCertificate",
        "parameters" : [ {
          "name" : "domainCode",
          "in" : "path",
          "description" : "Domain Id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "certificateCode",
          "in" : "path",
          "description" : "Certificate Id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Operation successfully done.",
            "schema" : {
              "$ref" : "#/definitions/IncapsulaResponse"
            }
          },
          "400" : {
            "description" : "Invalid data supplied"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
    "/tools/incapsula/sites/{domainCode}/configure/acl" : {
      "put" : {
        "tags" : [ "toolsincapsulasites" ],
        "summary" : "Configure acl in Incapsula",
        "description" : "Configure acl in Incapsula",
        "operationId" : "configureAcl",
        "parameters" : [ {
          "name" : "domainCode",
          "in" : "path",
          "description" : "Domain Id",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Acls Configuration",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/WAF AclsConfiguration"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Operation successfully done.",
            "schema" : {
              "$ref" : "#/definitions/IncapsulaResponse"
            }
          },
          "400" : {
            "description" : "Invalid data supplied"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
    "/tools/incapsula/sites/{domainCode}/configure/security" : {
      "put" : {
        "tags" : [ "toolsincapsulasites" ],
        "summary" : "Configure security in Incapsula",
        "description" : "Configure security in Incapsula",
        "operationId" : "configureSecurity",
        "parameters" : [ {
          "name" : "domainCode",
          "in" : "path",
          "description" : "Domain Id",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Threats Configuration",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/WAF ThreatsConfiguration"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Operation successfully done.",
            "schema" : {
              "$ref" : "#/definitions/IncapsulaResponse"
            }
          },
          "400" : {
            "description" : "Invalid data supplied"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
...
    "/tools/qualys/profiles/search" : {
      "post" : {
        "tags" : [ "toolsqualysprofiles" ],
        "summary" : "Search among all 'OptionProfile'",
        "description" : "Allows to search among all 'OptionProfile'",
        "operationId" : "getOptionProfiles",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Filters",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Filters"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Results page you want to retrieve (0..N), default value = O",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "Number of records per page, default value = 20",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "required" : false,
          "type" : "array",
          "items" : {
            "type" : "string"
          },
          "collectionFormat" : "multi"
        } ],
        "responses" : {
          "200" : {
            "description" : "OptionProfiles successfully returned.",
            "schema" : {
              "$ref" : "#/definitions/OptionProfile"
            }
          },
          "400" : {
            "description" : "Invalid data supplied"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
...
  },
  "definitions" : {
...
    "Criteria" : {
      "type" : "object",
      "properties" : {
        "field" : {
          "type" : "string"
        },
        "operator" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        }
      },
      "description" : "Criteria Object"
    },
...
    "Filters" : {
      "type" : "object",
      "properties" : {
        "Criteria" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Criteria"
          }
        }
      },
      "description" : "Filters Object"
    },
...
    "IncapsulaResponse" : {
      "type" : "object",
      "properties" : {
        "res" : {
          "type" : "integer",
          "format" : "int32"
        },
        "res_message" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string"
        },
        "ips" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "dns" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF Dn"
          }
        },
        "original_dns" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF OriginalDn"
          }
        },
        "warnings" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF Warning"
          }
        },
        "security" : {
          "$ref" : "#/definitions/WAF Security"
        },
        "active" : {
          "type" : "string"
        },
        "acceleration_level" : {
          "type" : "string"
        },
        "site_creation_date" : {
          "type" : "integer",
          "format" : "int64"
        },
        "sealLocation" : {
          "$ref" : "#/definitions/WAF SealLocation"
        },
        "ssl" : {
          "$ref" : "#/definitions/WAF Ssl"
        },
        "login_protect" : {
          "$ref" : "#/definitions/WAF LoginProtect"
        },
        "performance_configuration" : {
          "$ref" : "#/definitions/WAF PerformanceConfiguration"
        },
        "visits_timeseries" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF VisitsTimeseries"
          }
        },
        "requests_geo_dist_summary" : {
          "$ref" : "#/definitions/WAF RequestsGeoDistSummary"
        },
        "caching" : {
          "$ref" : "#/definitions/WAF Caching"
        },
        "caching_timeseries" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF CachingTimeseries"
          }
        },
        "hits_timeseries" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF HitsTimeseries"
          }
        },
        "threats" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF Threat"
          }
        },
        "visits_dist_summary" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF VisitsDistSummary"
          }
        },
        "bandwidth_timeseries" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF BandwidthTimeseries"
          }
        },
        "visits" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF Visit"
          }
        }
      },
      "description" : "IncapsulaResponse Object"
    },
...
    "OptionProfile" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "name" : {
          "type" : "string"
        },
        "owner" : {
          "$ref" : "#/definitions/Qualys User"
        },
        "isDefault" : {
          "type" : "boolean"
        },
        "timeoutErrorThreshold" : {
          "type" : "integer",
          "format" : "int64"
        },
        "unexpectedErrorThreshold" : {
          "type" : "integer",
          "format" : "int64"
        }
      },
      "description" : "OptionProfile Object"
    },
...
    "WAF" : {
      "type" : "object",
      "properties" : {
        "rules" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WAF Rules"
          }
        }
      },
      "description" : "WAF Object"
    },
...
  }
}